+++ app-backend/backsplash-core/src/main/scala/com/rasterfoundry/backsplash/BacksplashImage.scala
@@ -28,6 +28,7 @@ import scalacache.modes.sync._
   * need to be flagged with the @cacheKeyExclude decorator to avoid unecessarily namespacing values in the keys
   *
   * @param imageId UUID of the image (scene) in the database
+  * @param projectId UUID of the layer this image is a part of
   * @param uri location of the source data
   * @param footprint extent of data the image covers
   * @param subsetBands subset of bands to be read from source
@@ -35,6 +36,7 @@ import scalacache.modes.sync._
   * @param singleBandOptions band + options of how to color a single band
   */
 case class BacksplashImage(imageId: UUID,
+                           projectId: UUID,
                            uri: String,
                            @cacheKeyExclude footprint: MultiPolygon,
                            subsetBands: List[Int],